Replace stripe patttern with converging cards
